requirements.txt
Flask==3.1.1
Flask-Sock==0.7.0
gunicorn==23.0.0
kafka-python==2.2.15
neo4j==5.28.1
pg8000==1.31.5
prometheus-client==0.22.1
PyYAML==6.0.2
redis==6.4.0
